+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -6,6 +6,7 @@
   ([#2871](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2871))
 
 * Fixed the default endpoint for OTLP exporter over HTTP/Protobuf.
+  The default value is `http://localhost:4318`.
   ([#2868](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2868))
 
 ## 1.2.0-rc2